build(CMakeLists.txt): add OpenMP support to the project refactor(main.cpp): change function name from largestPathValue to numEnclaves The CMakeLists.txt file has been modified to add support for OpenMP to the project. This will allow the project to take advantage of multi-threading capabilities. The main.cpp file has been modified to change the function name from largestPathValue to numEnclaves. This is a more descriptive name for the function and improves code readability.
feat(Graph.h): add support for largestPathValue function
The largestPathValue function takes a string of colors and a vector of edges and returns the largest number of colors that can be visited in a directed graph. The function uses a depth-first search algorithm to traverse the graph and update the count of colors visited. The merge and update functions are used to update the count of colors visited. The function graph_dfs_acyclic is used to check if the graph is acyclic. The function canFinish and findOrder are modified to use the graph_dfs_acyclic function. A new test case is added to test the largestPathValue function.