feat: add solutions for several LeetCode problems Add solutions for the following LeetCode problems: - profitableSchemes - minInsertions - widthOfBinaryTree - longestZigZag
The profitableSchemes function calculates the number of profitable schemes given a number of people, minimum profit, and a list of groups and profits. The minInsertions function calculates the minimum number of insertions required to make a string a palindrome. The widthOfBinaryTree function calculates the maximum width of a binary tree. The longestZigZag function calculates the length of the longest zigzag path in a binary tree.

feat(DP.h): add function numberOfArrays to namespace dp::split_array
feat(test_dp.cpp): add test for function numberOfArrays
The function numberOfArrays is added to the dp::split_array namespace. It solves the problem of restoring an array from a string representation and returns the number of possible arrays that can be formed. A test is added to test_dp.cpp to verify the correctness of the implementation.